refactor: move prisma client from shop folder
