+++ src/lib/server/shop/addToCart/reservations.ts
@@ -1,4 +1,4 @@
-import authorizedPrismaClient from "$lib/server/shop/authorizedPrisma";
+import authorizedPrismaClient from "$lib/server/authorizedPrisma";
 import {
   ensurePaymentIntentState,
   removePaymentIntent,
