+++ src/routes/(app)/admin/debug/+page.server.ts
@@ -1,7 +1,7 @@
 import { env } from "$env/dynamic/private";
 import keycloak from "$lib/server/keycloak";
 import meilisearchSync from "$lib/search/sync";
-import authorizedPrismaClient from "$lib/server/shop/authorizedPrisma";
+import authorizedPrismaClient from "$lib/server/authorizedPrisma";
 import { isNollningPeriod } from "$lib/utils/adminSettings/nollning";
 
 export const load = async () => {
