+++ src/routes/(app)/expenses/upload/+page.server.ts
@@ -1,7 +1,7 @@
 import { PUBLIC_BUCKETS_FILES } from "$env/static/public";
 import { removeFilesWithoutAccessCheck } from "$lib/files/fileHandler";
 import { uploadFile } from "$lib/files/uploadFiles";
-import authorizedPrismaClient from "$lib/server/shop/authorizedPrisma";
+import authorizedPrismaClient from "$lib/server/authorizedPrisma";
 import { redirect } from "$lib/utils/redirect";
 import * as m from "$paraglide/messages";
 import type { Prisma } from "@prisma/client";
